Extract RegionConatiner
