+++ src/App.jsx
@@ -1,20 +1,16 @@
 import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { setSelectedCategory, setSelectedRegion } from './actions';
+import { setSelectedCategory } from './actions';
 import { get } from './utils';
+import RegionContainer from './RegionContainer';
 
 export default function App() {
   const dispatch = useDispatch();
 
-  const regions = useSelector(get('regions'));
   const selectedRegion = useSelector(get('selectedRegion'));
   const categories = useSelector(get('categories'));
   const selectedCategory = useSelector(get('selectedCategory'));
   const restaurants = useSelector(get('restaurants'));
-
-  const handleRegionClick = (name) => {
-    dispatch(setSelectedRegion(name));
-  };
 
   const handleCategoryClick = (name) => {
     dispatch(setSelectedCategory(name));
@@ -22,15 +18,7 @@ export default function App() {
 
   return (
     <>
-      <ul>
-        {regions.map((region) => (
-          <li key={region.id}>
-            <button type="button" onClick={() => handleRegionClick(region.name)}>
-              {region.name}
-            </button>
-          </li>
-        ))}
-      </ul>
+      <RegionContainer />
       <ul>
         {categories.map((category) => (
           <li key={category.id}>
