+++ src/App.test.jsx
@@ -16,6 +16,7 @@ test('App', () => {
     regions: [{ id: 1, name: '서울' }],
     categories: [{ id: 1, name: '한식' }],
   }));
+
   const { container } = render(
     <App />,
   );
